Fixed midnight entry
If a shot starts at midnight, the timestamp should read "0000".  If it
ends at midnight, the timestamp should read "2400".  It was always set
to "0000".  This fix added logic to the case statements for each method
to evaluate the endTime.  If equals 0, set it to 2400.  There will never
be an instance when an endTime should equal 0.