Fixed shots around midnight
Shots that end at midnight now have an end time of '2400' instead of
'0000' and are processed within the correct day.  Shots that start at
midnight are now processed on the correct day.  However, shots that
start at midnight on Monday are not showing up.  Flagged this issue with
FIXME in the code for later work.