Update highlighting added
Modified ScheduleGenieMenu.java to present a dialog to the operator
asking if this is a new or updated schedule.

Added OpenWorkbooks method to handle the workbook objects.

Added FileCompare method to verify each cell against the old schedule
and highlight any cells that are different.  Two issues currently exist
with this functionality:

	Issue 1:  If splitting a shot into two, the second shot does not get
highlighted.

	Issue 2:  If combining a shot that encompasses more than one line,
highlighting will be wrong for the rest of the schedule.

Neither issue affects the contents of the cells.  It is only a
highlighting issue.

Added unit testing to assist with effort.  Extended test of new methods
needs asserts.

Added new formatting methods in SpreadsheetHandler.java to assist with
highlighting.